refactor: modernize infrastructure with construct-based architecture
- Refactor CDK stacks to use construct pattern with proper separation of concerns
- Consolidate dev/prod environments into single unified stack
- Fix Lambda code path to use backend/app directory structure
- Add Lambda layers for FastAPI dependencies with proper grant methods
- Update Taskfile to use simplified commands for unified stack deployment
- Consolidate documentation under docs/ directory
- Remove legacy stack files and outdated deployment configurations
- Add proper Lambda handler with Mangum for FastAPI integration
- Implement IAM permissions using CDK grant methods for better security

Breaking Changes:
- Stack names changed from separate dev/prod to unified ShirayukiTomoFansiteStack
- CI/CD workflows need to be moved from infrastructure/.github to root .github directory (manual step required)
- Documentation structure reorganized under docs/ directory

Note: GitHub Actions workflow file needs to be manually moved from infrastructure/.github/workflows/ci.yml to .github/workflows/ci.yml due to token permissions.
