--- infrastructure/.github/workflows/ci.yml
@@ -1,164 +1,0 @@
-name: CI/CD Pipeline
-
-on:
-  push:
-    branches: [ main, develop ]
-    paths:
-      - 'infrastructure/**'
-  pull_request:
-    branches: [ main ]
-    paths:
-      - 'infrastructure/**'
-
-jobs:
-  test:
-    runs-on: ubuntu-latest
-    defaults:
-      run:
-        working-directory: infrastructure
-    
-    steps:
-    - uses: actions/checkout@v4
-    
-    - name: Set up Python
-      uses: actions/setup-python@v5
-      with:
-        python-version: '3.13'
-    
-    - name: Install uv
-      uses: astral-sh/setup-uv@v4
-      with:
-        version: "latest"
-    
-    - name: Install dependencies
-      run: uv sync --dev
-    
-    - name: Run linting
-      run: uv run ruff check .
-    
-    - name: Run type checking
-      run: uv run mypy .
-    
-    - name: Run tests
-      run: uv run pytest --cov=stacks --cov-report=xml
-    
-    - name: Upload coverage to Codecov
-      uses: codecov/codecov-action@v4
-      with:
-        file: ./infrastructure/coverage.xml
-        flags: infrastructure
-        name: infrastructure-coverage
-    
-    - name: Install Node.js
-      uses: actions/setup-node@v4
-      with:
-        node-version: '20'
-    
-    - name: Install AWS CDK
-      run: npm install -g aws-cdk
-    
-    - name: CDK Synth
-      run: uv run cdk synth
-    
-    - name: CDK Diff (Dev)
-      if: github.event_name == 'pull_request'
-      run: uv run cdk diff ShirayukiTomoFansiteDevStack
-      env:
-        AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID }}
-        AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
-        AWS_DEFAULT_REGION: ap-northeast-1
-      continue-on-error: true
-
-  deploy-dev:
-    if: github.ref == 'refs/heads/develop' && github.event_name == 'push'
-    needs: test
-    runs-on: ubuntu-latest
-    defaults:
-      run:
-        working-directory: infrastructure
-    environment: development
-    
-    steps:
-    - uses: actions/checkout@v4
-    
-    - name: Set up Python
-      uses: actions/setup-python@v5
-      with:
-        python-version: '3.13'
-    
-    - name: Install uv
-      uses: astral-sh/setup-uv@v4
-      with:
-        version: "latest"
-    
-    - name: Install dependencies
-      run: uv sync
-    
-    - name: Install Node.js
-      uses: actions/setup-node@v4
-      with:
-        node-version: '20'
-    
-    - name: Install AWS CDK
-      run: npm install -g aws-cdk
-    
-    - name: Configure AWS credentials
-      uses: aws-actions/configure-aws-credentials@v4
-      with:
-        aws-access-key-id: ${{ secrets.AWS_ACCESS_KEY_ID }}
-        aws-secret-access-key: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
-        aws-region: ap-northeast-1
-    
-    - name: CDK Bootstrap (if needed)
-      run: uv run cdk bootstrap
-      continue-on-error: true
-    
-    - name: CDK Deploy Dev
-      run: uv run cdk deploy ShirayukiTomoFansiteDevStack --require-approval never
-
-  deploy-prod:
-    if: github.ref == 'refs/heads/main' && github.event_name == 'push'
-    needs: test
-    runs-on: ubuntu-latest
-    defaults:
-      run:
-        working-directory: infrastructure
-    environment: production
-    
-    steps:
-    - uses: actions/checkout@v4
-    
-    - name: Set up Python
-      uses: actions/setup-python@v5
-      with:
-        python-version: '3.13'
-    
-    - name: Install uv
-      uses: astral-sh/setup-uv@v4
-      with:
-        version: "latest"
-    
-    - name: Install dependencies
-      run: uv sync
-    
-    - name: Install Node.js
-      uses: actions/setup-node@v4
-      with:
-        node-version: '20'
-    
-    - name: Install AWS CDK
-      run: npm install -g aws-cdk
-    
-    - name: Configure AWS credentials
-      uses: aws-actions/configure-aws-credentials@v4
-      with:
-        aws-access-key-id: ${{ secrets.AWS_ACCESS_KEY_ID }}
-        aws-secret-access-key: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
-        aws-region: ap-northeast-1
-    
-    - name: CDK Bootstrap (if needed)
-      run: uv run cdk bootstrap
-      continue-on-error: true
-    
-    - name: CDK Deploy Prod
-      run: uv run cdk deploy ShirayukiTomoFansiteProdStack --require-approval never
